requirements.txt
cachetools==5.3.1
cdifflib==1.2.6
colorama==0.4.6
google-auth==2.22.0
google-auth-oauthlib==1.0.0
gspread==5.10.0
numpy==1.25.1
oauthlib==3.2.2
pandas==2.0.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pytz==2023.3
requests-oauthlib==1.3.1
rsa==4.9
statistics==1.0.3.5
tzdata==2023.3
wonderwords==2.2.0
